requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.1.1
django-redis==5.4.0
hiredis==3.0.0
idna==3.10
mysqlclient==2.2.4
O365==2.0.36
oauthlib==3.2.2
pillow==10.4.0
python-dateutil==2.9.0.post0
python-decouple==3.8
redis==5.2.0
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
